split admin mcp endpoints into submodules
